renamed favicon logo file and there usage
