+++ components/centraldashboard/public/components/main-page.js
@@ -28,7 +28,7 @@ import {html, PolymerElement} from '@polymer/polymer/polymer-element.js';
 
 import css from './main-page.css';
 import template from './main-page.pug';
-import logo from '../assets/logo3.svg';
+import logo from '../assets/logo.svg';
 
 import './registration-page.js';
 import './namespace-selector.js';
