+++ components/centraldashboard/public/index.html
@@ -10,7 +10,7 @@
 
     <title>Autonomize Model Hub</title>
 
-    <link rel="shortcut icon" href="/assets/autonomize-logo.png">
+    <link rel="shortcut icon" href="/assets/favicon.png">
     <!-- <link rel="icon" href="/assets/favicon-32x32.png" sizes="32x32">
     <link rel="icon" href="/assets/favicon-57x57.png" sizes="57x57">
     <link rel="icon" href="/assets/favicon-76x76.png" sizes="76x76">
